renamed test_small_weights to test_zero_weights
